activeEvent のディレクトリを削除
